Implement hMonitorVars as a local binding
Motivation for this is that 3 less parameters have to be passed to
hMonitorVars by hRunHsStatus and itself. All but the previous value are
pass to or declared in hRunHsStatus, so I'm naively guessing it would be
better to define the function there, using those bindings, than to pass
them as parameters.

Very uncertain of the effect this has on performance (if any at all).