Use iNotifyWatch for battery field
Instead of going based on an interval, use inotify to trigger the
battery field to update. This, in theory, should be more efficient than
just going on a set interval.

I tried just checking for modifications to status and now, but they do
not broadcast any events on my laptop. To get around this, it watches
uevent. I'm not sure if this is efficient or portable, but it works for
me.