Vastly simplify fields and main
Fields are now simply just pairs of starting values and functions that
take a semaphore (to trigger a print) and a variable (to write results
to). They are responsible for forking themselves if needed. This lets
threads not fork themselves if need.

The aforementioned result/state variables are now not wrapped in Either
ByteString. They are now also stored in individual variables instead of
tuples. Updates are sent to them by writing to them directly, instead
of the queues of lambdas from before. This significantly more simple
and should allow for more flexibility as well.

Startup now simply initializes fields, collecting their variables and
any spawned thread ids; forks a thread to print watch the print
semaphore; installs signal handlers; and waits for process termination.

FieldUtils and Fields/* were modified to accomodate the above changes.
It's not pretty. At all. They will need to be rewritten.