Simplify brightnessMonitor
brightnessMonitor was rewritten. Exception handling was removed. INotify now
watches for CloseWrite, since from my testing, Modify appears to be called
twice when the brightness is changed, but CloseWrite is only called once.

A "lock" was added so only one thread can access the brightness file at a time.
Not sure if this is needed, but it looks like it could prevent two getLines
before it was seeked back the beginning.

Instead of using custom types for specifying paths, just plain Strings and Ints
are used.