Replace semaphore and IO refs with chan and vects
Each fields' state can only be a ByteString.

Format/output function is removed; all states are appended together
and printed.

State is maintained within an immutable vector (but is updated
destructively) as opposed to several IORefs.

State is updated by writing the new ByteString to the channel, paired
with an index.

Printing is triggered when a new state is written to the channel.